Python 2.7 compatibility - math.floor returns float in 2.7 and range requires an int for the step size. In Python 3.x it returns an int. I know that it will always be an int, so I cast it to an int.

+++ matrixprofile/scrimp.py
@@ -76,7 +76,7 @@ def calc_exclusion_zone(window_size):
 
 
 def calc_step_size(window_size, step_size):
-    return math.floor(window_size * step_size)
+    return int(math.floor(window_size * step_size))
 
 
 def calc_profile_len(n, window_size):
